#15 Separate out doughnut and line data, add horizontal margin and better copy
