+++ src/components/Infograph.js
@@ -1,49 +1,22 @@
-import { Doughnut } from 'react-chartjs-2';
-import { HorizontalBar } from 'react-chartjs-2';
+import { Doughnut, HorizontalBar } from 'react-chartjs-2';
 import React from 'react';
-
-const doughnutData = {
-  labels: ['Red', 'Green', 'Yellow'],
-  datasets: [
-    {
-      data: [300, 50, 100],
-      backgroundColor: ['#FF6384', '#36A2EB', '#FFCE56'],
-      hoverBackgroundColor: ['#FF6384', '#36A2EB', '#FFCE56']
-    }
-  ]
-};
-
-const horizontalData = {
-  labels: ['January', 'February', 'March', 'April', 'May', 'June', 'July'],
-  datasets: [
-    {
-      label: 'My First dataset',
-      backgroundColor: 'rgba(255,99,132,0.2)',
-      borderColor: 'rgba(255,99,132,1)',
-      borderWidth: 1,
-      hoverBackgroundColor: 'rgba(255,99,132,0.4)',
-      hoverBorderColor: 'rgba(255,99,132,1)',
-      data: [65, 59, 80, 81, 56, 55, 40]
-    }
-  ]
-};
+import doughnutData from '../data/doughnut';
+import horizontalData from '../data/horizontal';
 
 const Infograph = () => {
   return (
     <div>
-      <h1 className="info-title tc f4">First section</h1>
+      <h1 className="info-title tc f4">Introduction</h1>
       <p className="info-desc tc f6">
-        More info on this piece of info will go here. More info on this piece of
-        info will go here. More info on this piece of info will go here.
+        More info on this piece of info will go here.
       </p>
-      <h1 className="info-title tc f4">Second section</h1>
+      <h1 className="info-title tc f4">Graph 1</h1>
       <Doughnut data={doughnutData} width={100} height={50} />
-      <h1 className="info-title tc f4">Third section</h1>
       <p className="info-desc tc f6">
         More info on this piece of info will go here. More info on this piece of
         info will go here. More info on this piece of info will go here.
       </p>
-      <h1 className="info-title tc f4">Fourth section</h1>
+      <h1 className="info-title tc f4">Graph 2</h1>
       <HorizontalBar data={horizontalData} />
     </div>
   );
