+++ src/data/doughnut.js
@@ -1,0 +1,12 @@
+const doughnutData = {
+  labels: ['Scotland', 'England', 'Wales', 'Northern Ireland'],
+  datasets: [
+    {
+      data: [300, 50, 100, 70],
+      backgroundColor: ['#36A2EB', '#FF6384', '#FFCE56', '#33e75a'],
+      hoverBackgroundColor: ['#36A2EB', '#FF6384', '#FFCE56', '#33e75a']
+    }
+  ]
+};
+
+export default doughnutData;
